Remove graph construction primitives from the internal module
